% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comtrade.R
\name{comtrade.data}
\alias{comtrade.data}
\title{comtrade.data}
\usage{
comtrade.data(r, p, dates = "now", freq = "A", type = "C", px = "HS",
  rg = "all", cc = "TOTAL", maxrec = 50000, validation = FALSE)
}
\arguments{
\item{r}{reporting area - either "all" or a numeric vector of reporting area codes, see \code{\link{comtrade.codes}}}

\item{p}{partner area - either "all" or a numeric vector of partner area codes, see \code{\link{comtrade.codes}}}

\item{dates}{time periods - "now" (default), "recent" (5 most recent years/months), or a list of dates, strings or integers
indicating which years/months to include. The strings/integers should be in the form "YYYY" or "YYYYMM" where YYYY=year
and MM=month}

\item{freq}{frequency - "A"=annual (default), "M"=monthly}

\item{type}{type of trade - "C"=commodities (default), "S"=services (not yet available)}

\item{px}{classification - "HS"=harmonized system (default), "H0"=HS 1992, "H1"=HS 1996, "H2"=HS 2002, "H3"=HS 2007,
"H4"=HS 2012, "ST"=Standard International Trade Classification (SITC), "S1"=SITC revision 1, "S2"=SITC revision 2,
"S3"=SITC revision 3, "S4"=SITC revision 4, "BEC"=Broad Economic Categories.}

\item{rg}{trade regime/flow - "all" (default), "1" (imports), "2" (exports), "3" (re-Export), "4" (re-Import)}

\item{cc}{classification code - "AG1", "AG2" (default), "AG3", "AG4", "AG5", "AG6", "TOTAL" or "ALL"}

\item{maxrec}{maximum number of records to download (default = 50000 the maximum possible number)}

\item{validation}{whether or not to return validation information (showing download status, time, etc.). Default is FALSE.}
}
\value{
A list containing a validation attribute (NULL for csv output), and a data attribute containing the data
}
\description{
Download data from http://comtrade.un.org/data/
}
\details{
For a more detailed explanation of the parameters see: http://comtrade.un.org/data/doc/api/
}
\author{
Ben Veal
}

