% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comtrade.R
\name{comtrade.codes}
\alias{comtrade.codes}
\title{comtrade.codes}
\usage{
comtrade.codes()
}
\value{
A data.frame containing country names and associated reporter numbers.
}
\description{
Show country codes for comtrade.data reporter and partner params
}
\details{
This function is copied directly from http://comtrade.un.org/data/Doc/api/ex/r and works as of 06/05/2015
}
\author{
Ben Veal
}

