% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comtrade.R
\name{comtrade.data}
\alias{comtrade.data}
\title{comtrade.data}
\usage{
comtrade.data(url = "http://comtrade.un.org/api/get?", maxrec = 50000,
  type = "C", freq = "A", px = "HS", ps = "now", r, p, rg = "all",
  cc = "TOTAL", fmt = "json")
}
\arguments{
\item{url}{url for http api (you probably don't need to touch this)}

\item{maxrec}{maximum number of records to download (default = 50000 the maximum possible number)}

\item{type}{type of trade - "C" (default), "S" (not yet available) trade in services data}

\item{freq}{frequency - "A"=annual (default), "M"=monthly}

\item{px}{classification - "HS"=harmonized system (default), "H0"=HS 1992, "H1"=HS 1996, "H2"=HS 2002, "H3"=HS 2007,
"H4"=HS 2012, "ST"=Standard International Trade Classification (SITC), "S1"=SITC revision 1, "S2"=SITC revision 2,
"S3"=SITC revision 3, "S4"=SITC revision 4, "BEC"=Broad Economic Categories.}

\item{ps}{time period - "now" (default), "recent", or "YYYY" or "YYYYMM" where YYYY=year and MM=month}

\item{r}{reporting area - "all", see \code{\link{comtrade.codes}}}

\item{p}{partner area - "all", see \code{\link{comtrade.codes}}}

\item{rg}{trade regime/flow - "all" (default), "1" (imports), "2" (exports), "3" (re-Export), "4" (re-Import)}

\item{cc}{classification code - "AG1", "AG2" (default), "AG3", "AG4", "AG5", "AG6", "TOTAL" or "ALL"}

\item{fmt}{output format - "json" (default), or "csv". json output includes validation information.}
}
\value{
A list containing a validation attribute (NULL for csv output), and a data attribute containing the data
}
\description{
Download data from http://comtrade.un.org/data/
}
\details{
This function is copied directly from http://comtrade.un.org/data/Doc/api/ex/r and works as of 06/05/2015
To see a more detailed explanation of the parameters see: http://comtrade.un.org/data/doc/api/
}
\author{
Ben Veal
}

